hayro-common: Add support for any bit size between 1 and 32 for bit writer (#504)

+++ hayro-jpeg2000/src/arithmetic_decoder.rs
@@ -295,7 +295,7 @@ mod tests {
 
         for _ in 0..expected_output.len() {
             for _ in 0..8 {
-                let next = decoder.decode(&mut ctx) as u16;
+                let next = decoder.decode(&mut ctx);
                 writer.write(next);
             }
         }
